@@ -1,6 +1,6 @@
 name: mapbox_api
 description: Unofficial Mapbox Dart SDK that follow official Mapbox documentation.
-version: 0.3.0
+version: 1.0.0
 homepage: https://github.com/stafyniaksacha/mapbox-api
 
 dependencies:
@@ -10,4 +10,4 @@ dev_dependencies:
   pedantic: ^1.9.2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
