@@ -53,10 +53,10 @@ void main() async {
     return;
   }
 
-  if (response.routes.isNotEmpty) {
-    final route = response.routes[0];
+  if (response.routes!.isNotEmpty) {
+    final route = response.routes![0];
     final eta = Duration(
-      seconds: route.duration.toInt(),
+      seconds: route.duration!.toInt(),
     );
     print(eta.toString());
   }
